+++ packages/components/src/components.d.ts
@@ -306,6 +306,10 @@ export namespace Components {
          */
         "iconBadgeValue": string;
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level": string;
+        /**
           * @description Location object.
          */
         "location": any;
@@ -878,6 +882,10 @@ export namespace Components {
           * Id for the input field.
          */
         "idAttribute": string;
+        /**
+          * The language used when retrieving Google Places or Mapbox autocomplete suggestions.
+         */
+        "language": string;
         /**
           * If searching should include Mapbox autocomplete suggestions.
          */
@@ -1670,6 +1678,10 @@ declare namespace LocalJSX {
          */
         "iconBadgeValue"?: string;
         /**
+          * @description The word used for "Level" when showing level info. Default is "Level".
+         */
+        "level"?: string;
+        /**
           * @description Location object.
          */
         "location"?: any;
@@ -2214,6 +2226,10 @@ declare namespace LocalJSX {
           * Id for the input field.
          */
         "idAttribute"?: string;
+        /**
+          * The language used when retrieving Google Places or Mapbox autocomplete suggestions.
+         */
+        "language"?: string;
         /**
           * If searching should include Mapbox autocomplete suggestions.
          */