+++ packages/map-template/src/components/Accessibility/Accessibility.jsx
@@ -3,6 +3,7 @@ import Tooltip from '../Tooltip/Tooltip';
 import './Accessibility.scss';
 import primaryColorState from '../../atoms/primaryColorState';
 import accessibilityOnState from '../../atoms/accessibilityOnState';
+import { t } from 'i18next';
 
 /**
  * Accessibility component which uses the mi-toggle and displays an info tooltip.
@@ -15,8 +16,8 @@ function Accessibility() {
 
     return <div className="accessibility">
         <input className="mi-toggle" type="checkbox" checked={accessibilityOn} onChange={e => setAccessibilityOn(e.target.checked)} style={{ backgroundColor: accessibilityOn ? primaryColor : '' }} />
-        <div>Accessibility</div>
-        <Tooltip text="Turn on Accessibility to get directions that avoid stairs and escalators."></Tooltip>
+        <div>{t('Accessibility')}</div>
+        <Tooltip text={t('Turn on Accessibility to get directions that avoid stairs and escalators.')}></Tooltip>
     </div>
 }