+++ packages/map-template/src/i18n/initialize.js
@@ -1,0 +1,32 @@
+import i18n from 'i18next';
+import { initReactI18next } from 'react-i18next';
+import en from './en.js';
+import da from './da.js';
+import de from './de.js';
+import fr from './fr.js';
+
+export default function initI18n(language) {
+    i18n
+        .use(initReactI18next)
+        .init({
+            resources: {
+                en: {
+                    translation: en
+                },
+                da: {
+                    translation: da
+                },
+                de: {
+                    translation: de
+                },
+                fr: {
+                    translation: fr
+                }
+            },
+            lng: language,
+            fallbackLng: 'en',
+            interpolation: {
+                escapeValue: false // react already safes from xss => https://www.i18next.com/translation-function/interpolation#unescape
+            }
+        });
+}