+++ packages/components/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.14.1] - 2023-12-20
+
+### Fixed
+
+- The `mi-search` component that can now get external location info in any language by using the `language` property.
+
 ## [13.14.0] - 2023-11-06
 
 ### Added
@@ -16,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add support for searching for Mapbox places in the `mi-search` component. 
+- Add support for searching for Mapbox places in the `mi-search` component.
 
 ## [13.12.0] - 2023-10-06